+++ src/js/utils/general.ts
@@ -1,11 +1,12 @@
+import { type Step } from '../step';
 import { isFunction, isString } from './type-check';
 
 /**
  * Ensure class prefix ends in `-`
- * @param {string} prefix The prefix to prepend to the class names generated by nano-css
- * @return {string} The prefix ending in `-`
+ * @param prefix - The prefix to prepend to the class names generated by nano-css
+ * @return The prefix ending in `-`
  */
-export function normalizePrefix(prefix) {
+export function normalizePrefix(prefix: string) {
   if (!isString(prefix) || prefix === '') {
     return '';
   }
@@ -15,12 +16,12 @@ export function normalizePrefix(prefix) {
 
 /**
  * Resolves attachTo options, converting element option value to a qualified HTMLElement.
- * @param {Step} step The step instance
+ * @param step - The step instance
  * @returns {{}|{element, on}}
  * `element` is a qualified HTML Element
  * `on` is a string position value
  */
-export function parseAttachTo(step) {
+export function parseAttachTo(step: Step) {
   const options = step.options.attachTo || {};
   const returnOpts = Object.assign({}, options);
 
@@ -66,7 +67,6 @@ export function shouldCenterStep(resolvedAttachToOptions) {
 
 /**
  * Create a unique id for steps, tours, modals, etc
- * @return {string}
  */
 export function uuid() {
   let d = Date.now();
