+++ src/js/utils/type-check.ts
@@ -1,0 +1,39 @@
+/**
+ * Checks if `value` is classified as an `Element`.
+ * @param value The param to check if it is an Element
+ */
+export function isElement(value: unknown) {
+  return value instanceof Element;
+}
+
+/**
+ * Checks if `value` is classified as an `HTMLElement`.
+ * @param value The param to check if it is an HTMLElement
+ */
+export function isHTMLElement(value: unknown) {
+  return value instanceof HTMLElement;
+}
+
+/**
+ * Checks if `value` is classified as a `Function` object.
+ * @param value The param to check if it is a function
+ */
+export function isFunction(value: unknown) {
+  return typeof value === 'function';
+}
+
+/**
+ * Checks if `value` is classified as a `String` object.
+ * @param value The param to check if it is a string
+ */
+export function isString(value: unknown) {
+  return typeof value === 'string';
+}
+
+/**
+ * Checks if `value` is undefined.
+ * @param value The param to check if it is undefined
+ */
+export function isUndefined(value: unknown) {
+  return value === undefined;
+}
